Move JS config and JS files to their proper locations
- jQuery FlexSlider's original JS and CSS files now live in vendor/assets as
they should (see http://guides.rubyonrails.org/asset_pipeline.html:
"vendor/assets is for assets that are owned by outside entities, such as code
for JavaScript plugins and CSS frameworks.")
- This extension's JS configuration now lives in the partial rather than a JS
file. While it would be preferable to have it in a JS file, this poses problems
with configurations where the environment isn't initialized when assets are
precompiled (e.g. on Heroku).
- As an added bonus ;), comments for the JS configuration are now aligned
properly for better readability.